+++ tests/Tgstation.Server.Host.Tests/Security/TestAuthenticationContext.cs
@@ -44,8 +44,8 @@ namespace Tgstation.Server.Host.Security.Tests
 
 			user.AdministrationRights = AdministrationRights.EditUsers;
 			instanceUser.ByondRights = ByondRights.ChangeVersion | ByondRights.ReadActive;
-			Assert.AreEqual((int)user.AdministrationRights, authContext.GetRight(RightsType.Administration));
-			Assert.AreEqual((int)instanceUser.ByondRights, authContext.GetRight(RightsType.Byond));
+			Assert.AreEqual((ulong)user.AdministrationRights, authContext.GetRight(RightsType.Administration));
+			Assert.AreEqual((ulong)instanceUser.ByondRights, authContext.GetRight(RightsType.Byond));
 		}
 	}
 }
